# internal/server/gen/swagger.yaml
components:
  schemas:
    ent.AIInsight:
      properties:
        confidence:
          description: 置信度 0.0~1.0
          type: number
        content:
          additionalProperties: {}
          description: AI 产出内容
          type: object
        created_at:
          description: CreatedAt holds the value of the "created_at" field.
          type: string
        edges:
          $ref: '#/components/schemas/ent.AIInsightEdges'
        evidence:
          description: 依据，每条 AI 建议必须可溯源
          items:
            additionalProperties: {}
            type: object
          type: array
          uniqueItems: false
        id:
          description: ID of the ent.
          type: integer
        stage:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_aiinsight.Stage'
        status:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_aiinsight.Status'
        type:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_aiinsight.Type'
        updated_at:
          description: UpdatedAt holds the value of the "updated_at" field.
          type: string
      type: object
...
    ent.EscalationPolicy:
      description: EscalationPolicy holds the value of the escalation_policy edge.
      properties:
        created_at:
          description: CreatedAt holds the value of the "created_at" field.
          type: string
        edges:
          $ref: '#/components/schemas/ent.EscalationPolicyEdges'
        id:
          description: ID of the ent.
          type: integer
        levels:
          description: 有序升级层级
          items:
            $ref: '#/components/schemas/github_com_kevin_vigil_ent_schema.EscalationLevel'
          type: array
          uniqueItems: false
        name:
          description: Name holds the value of the "name" field.
          type: string
        repeat_times:
          description: RepeatTimes holds the value of the "repeat_times" field.
          type: integer
        updated_at:
          description: UpdatedAt holds the value of the "updated_at" field.
          type: string
      type: object
...
    ent.Event:
      properties:
        created_at:
          description: CreatedAt holds the value of the "created_at" field.
          type: string
        dedup_key:
          description: 去重键
          type: string
        detail:
          additionalProperties: {}
          description: 原始 payload 明细
          type: object
        edges:
          $ref: '#/components/schemas/ent.EventEdges'
        id:
          description: ID of the ent.
          type: integer
        is_noise:
          description: 分诊判定为噪音
          type: boolean
        labels:
          additionalProperties:
            type: string
          description: 路由用标签
          type: object
        received_at:
          description: ReceivedAt holds the value of the "received_at" field.
          type: string
        severity:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_event.Severity'
        source:
          description: 告警源，如 prometheus
          type: string
        source_event_id:
          description: 原始告警 ID，去重 + 幂等键
          type: string
        status:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_event.Status'
        summary:
          description: 一句话摘要
          type: string
      type: object
...
    ent.Incident:
      description: Incident holds the value of the incident edge.
      properties:
        acked_at:
          description: 确认时间，MTTA 计算
          type: string
        closed_at:
          description: ClosedAt holds the value of the "closed_at" field.
          type: string
        created_at:
          description: CreatedAt holds the value of the "created_at" field.
          type: string
        current_level:
          description: 当前升级层级
          type: integer
        edges:
          $ref: '#/components/schemas/ent.IncidentEdges'
        embedding:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_schema.NullableVector'
        escalated_count:
          description: 已升级次数
          type: integer
        id:
          description: ID of the ent.
          type: integer
        merged_into:
          description: 若被合并，指向主 Incident
          type: string
        number:
          description: 人类可读编号，如 INC-0042
          type: string
        priority:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_incident.Priority'
        resolved_at:
          description: ResolvedAt holds the value of the "resolved_at" field.
          type: string
        severity:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_incident.Severity'
        status:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_incident.Status'
        summary:
          description: 当前概要，可随处置更新
          type: string
        title:
          description: Title holds the value of the "title" field.
          type: string
        trigger_source_event_id:
          description: TriggerSourceEventID holds the value of the "trigger_source_event_id"
            field.
          type: string
        trigger_type:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_incident.TriggerType'
        updated_at:
          description: UpdatedAt holds the value of the "updated_at" field.
          type: string
        war_room:
          additionalProperties: {}
          description: 作战室：im_platform/im_channel_id/created_at
          type: object
      type: object
    ent.IncidentAction:
      properties:
        actor:
          additionalProperties:
            type: string
          description: 'actor: kind + id'
          type: object
        edges:
          $ref: '#/components/schemas/ent.IncidentActionEdges'
        id:
          description: ID of the ent.
          type: integer
        payload:
          additionalProperties: {}
          description: Payload holds the value of the "payload" field.
          type: object
        result:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_incidentaction.Result'
        timestamp:
          description: Timestamp holds the value of the "timestamp" field.
          type: string
        type:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_incidentaction.Type'
        via:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_incidentaction.Via'
      type: object
...
    ent.IncidentEdges:
      description: |-
        Edges holds the relations/edges for other nodes in the graph.
        The values are being populated by the IncidentQuery when eager-loading is set.
      properties:
        actions:
          description: Actions holds the value of the actions edge.
          items:
            $ref: '#/components/schemas/ent.IncidentAction'
          type: array
          uniqueItems: false
        ai_insights:
          description: AiInsights holds the value of the ai_insights edge.
          items:
            $ref: '#/components/schemas/ent.AIInsight'
          type: array
          uniqueItems: false
        assignee:
          $ref: '#/components/schemas/ent.User'
        escalation_policy:
          $ref: '#/components/schemas/ent.EscalationPolicy'
        events:
          description: Events holds the value of the events edge.
          items:
            $ref: '#/components/schemas/ent.Event'
          type: array
          uniqueItems: false
        postmortem:
          $ref: '#/components/schemas/ent.Postmortem'
        responders:
          description: Responders holds the value of the responders edge.
          items:
            $ref: '#/components/schemas/ent.User'
          type: array
          uniqueItems: false
        service:
          $ref: '#/components/schemas/ent.Service'
        team:
          $ref: '#/components/schemas/ent.Team'
        timeline:
          description: Timeline holds the value of the timeline edge.
          items:
            $ref: '#/components/schemas/ent.TimelineItem'
          type: array
          uniqueItems: false
      type: object
...
    ent.Postmortem:
      description: Postmortem holds the value of the postmortem edge.
      properties:
        created_at:
          description: CreatedAt holds the value of the "created_at" field.
          type: string
        edges:
          $ref: '#/components/schemas/ent.PostmortemEdges'
        embedding:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_schema.NullableVector'
        generated_by:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_postmortem.GeneratedBy'
        id:
          description: ID of the ent.
          type: integer
        published_at:
          description: PublishedAt holds the value of the "published_at" field.
          type: string
        sections:
          additionalProperties: {}
          description: 结构化内容章节
          type: object
        status:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_postmortem.Status'
        updated_at:
          description: UpdatedAt holds the value of the "updated_at" field.
          type: string
      type: object
...
    ent.User:
      description: User holds the value of the user edge.
      properties:
        created_at:
          description: CreatedAt holds the value of the "created_at" field.
          type: string
        edges:
          $ref: '#/components/schemas/ent.UserEdges'
        email:
          description: 邮箱
          type: string
        id:
          description: ID of the ent.
          type: integer
        im_accounts:
          description: IM 账号绑定（JSON，兼容字段；新数据见 im_bindings 表）
          items:
            $ref: '#/components/schemas/github_com_kevin_vigil_ent_schema.IMAccount'
          type: array
          uniqueItems: false
        name:
          description: 显示名
          type: string
        phone:
          description: 电话，用于 SMS/语音
          type: string
        status:
          $ref: '#/components/schemas/github_com_kevin_vigil_ent_user.Status'
        timezone:
          description: Timezone holds the value of the "timezone" field.
          type: string
        updated_at:
          description: UpdatedAt holds the value of the "updated_at" field.
          type: string
        username:
          description: 登录名
          type: string
      type: object
...
    github_com_kevin_vigil_ent_incident.Priority:
      description: Priority holds the value of the "priority" field.
      enum:
      - p3
      - p1
      - p2
      - p4
      type: string
      x-enum-varnames:
      - DefaultPriority
      - PriorityP1
      - PriorityP2
      - PriorityP3
      - PriorityP4
    github_com_kevin_vigil_ent_incident.Severity:
      description: Severity holds the value of the "severity" field.
      enum:
      - critical
      - warning
      - info
      type: string
      x-enum-varnames:
      - SeverityCritical
      - SeverityWarning
      - SeverityInfo
    github_com_kevin_vigil_ent_incident.Status:
      description: Status holds the value of the "status" field.
      enum:
      - triggered
      - escalated
      - acked
      - resolved
      - closed
      type: string
      x-enum-varnames:
      - DefaultStatus
      - StatusTriggered
      - StatusEscalated
      - StatusAcked
      - StatusResolved
      - StatusClosed
    github_com_kevin_vigil_ent_incident.TriggerType:
      description: TriggerType holds the value of the "trigger_type" field.
      enum:
      - auto
      - manual
      - merged
      type: string
      x-enum-varnames:
      - DefaultTriggerType
      - TriggerTypeAuto
      - TriggerTypeManual
      - TriggerTypeMerged
...
    github_com_kevin_vigil_ent_schema.NullableVector:
      description: 语义向量，published 复盘入库后计算，知识沉淀检索用
      properties:
        valid:
          type: boolean
      type: object
...
    github_com_kevin_vigil_internal_httputil.AckResponse:
      properties:
        raw_event_id:
          description: RawEventID ingestion 落库的 RawEvent ID（限接入 webhook）。
          example: 42
          type: integer
        retry_after:
          description: RetryAfter 限流/背压时建议重试秒数（限接入 webhook）。
          example: 60
          type: integer
        status:
          example: accepted
          type: string
      type: object
    github_com_kevin_vigil_internal_httputil.ErrorResponse:
      properties:
        code:
          description: Code 机器可读错误码（可选，非 4xx 通用错误可留空）。
          example: invalid_argument
          type: string
        details:
          description: Details 结构化补充信息（可选，如字段级校验错误）。
        error:
          example: invalid id
          type: string
      type: object
...
    github_com_kevin_vigil_internal_httputil.Paginated-ent_Incident:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ent.Incident'
          type: array
          uniqueItems: false
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: 业务 API 推荐（唯一声明）方案：HTTP Bearer JWT（/auth/login 换取）。
      scheme: bearer
      type: http
...
info:
  description: |-
    Vigil 告警处置平台 REST API。
    认证：业务 API 要求 Bearer JWT（/auth/login 换取）。鉴权中间件另接受 X-Vigil-User-ID 头作为本地/回退身份（可伪造，仅限受信网络，生产禁用），该回退方案不在 securitySchemes 中声明。
  license:
    name: MIT
  title: Vigil API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /incidents:
    get:
      description: 按状态/严重度过滤并分页返回事件，total 与筛选条件一致。
      parameters:
      - description: 按状态过滤
        in: query
        name: status
        schema:
          enum:
          - triggered
          - escalated
          - acked
          - resolved
          - closed
          type: string
      - description: 按严重度过滤
        in: query
        name: severity
        schema:
          enum:
          - critical
          - warning
          - info
          type: string
      - description: 分页大小（默认 50，上限 200）
        in: query
        name: limit
        schema:
          default: 50
          maximum: 200
          type: integer
      - description: 分页偏移
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.Paginated-ent_Incident'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: 查询事件列表
      tags:
      - incident
  /incidents/{id}:
    get:
      description: 含 responders 与 events 关联。
      parameters:
      - description: 事件 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ent.Incident'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Not Found
      security:
      - bearerAuth: []
      summary: 事件详情
      tags:
      - incident
  /incidents/{id}/ack:
    post:
      parameters:
      - description: 事件 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ent.Incident'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
      security:
      - bearerAuth: []
      summary: 确认事件（ack）
      tags:
      - incident
...
  /incidents/{id}/escalate:
    post:
      parameters:
      - description: 事件 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ent.Incident'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
      security:
      - bearerAuth: []
      summary: 升级事件（escalate）
      tags:
      - incident
...
  /incidents/{id}/resolve:
    post:
      parameters:
      - description: 事件 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ent.Incident'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
      security:
      - bearerAuth: []
      summary: 解决事件（resolve）
      tags:
      - incident
...
  /webhook/{token}:
    post:
      description: token 鉴权 → 限流/背压检查（payload 仍落库）→ 落 RawEvent → 入归一化队列 → 秒级返回 202。公开入口，不走
        RBAC。
      parameters:
      - description: 接入点鉴权 token（对应 Integration.token）
        in: path
        name: token
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.AckResponse'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Unauthorized
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.AckResponse'
          description: Too Many Requests
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.AckResponse'
          description: Service Unavailable
      summary: 接收告警 webhook
      tags:
      - ingestion
servers:
- url: /api/v1